allGameStates data separated from app.js to gameState.models
